+++ src/modules/admin/views/entry/create.php
@@ -7,6 +7,7 @@
  * @var \davidhirtz\yii2\cms\models\Entry $entry
  */
 
+use davidhirtz\yii2\cms\modules\admin\widgets\forms\EntryActiveForm;
 use davidhirtz\yii2\cms\modules\admin\widgets\navs\Submenu;
 use davidhirtz\yii2\skeleton\helpers\Html;
 use davidhirtz\yii2\skeleton\widgets\bootstrap\Panel;
@@ -22,7 +23,7 @@ $this->setTitle(Yii::t('cms', 'Create New Entry'));
 
 <?= Panel::widget([
     'title' => $this->title,
-    'content' => $entry->getActiveForm()::widget([
+    'content' => EntryActiveForm::widget([
         'model' => $entry,
     ]),
 ]); ?>